# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
security:
  - ClerkBearerAuth: []
...
paths:
...
  /nutrition/logs:
    get:
      operationId: nutritionListLogs
      tags: [nutrition]
      summary: List nutrition logs for the signed-in member
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        - name: to
          in: query
          required: false
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}-\\d{2}$"
      responses:
        "200":
          description: Nutrition logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NutritionLog"
        "400":
          description: Invalid date range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Member access is not approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /nutrition/logs/{date}:
    get:
      operationId: nutritionGetLog
      tags: [nutrition]
      summary: Get one nutrition log for the signed-in member
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}-\\d{2}$"
      responses:
        "200":
          description: Nutrition log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionLog"
        "400":
          description: Invalid date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Member access is not approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: nutritionPutLog
      tags: [nutrition]
      summary: Upsert one nutrition log for the signed-in member
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}-\\d{2}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertNutritionLogBody"
      responses:
        "200":
          description: Saved nutrition log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionLog"
        "400":
          description: Invalid nutrition log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Member access is not approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ClerkBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Clerk session bearer token sent as Authorization: Bearer <token>."
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]
      required:
        - error
...
    NutritionEntry:
      type: object
      additionalProperties: true
      properties:
        id:
          type: string
        foodId:
          type: string
        name:
          type: string
        mealType:
          type: string
        servings:
          type: number
        servingSize:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        fiber:
          type: number
        timestamp:
          type: number
        fromPhoto:
          type: boolean
        photoUri:
          type: string
      required:
        - id
        - name
    NutritionLog:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/NutritionEntry"
        waterIntake:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - date
        - entries
        - waterIntake
    UpsertNutritionLogBody:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/NutritionEntry"
        waterIntake:
          type: number
      required:
        - entries
        - waterIntake
...